code: doctor no longer optimizing WARN functionality changed
While documenting/reviewing I realized that the optimizing code couldn't
have any effect as otherwise it would have ended up calling
`createAndProcessScript` of the `AsyncScriptProcessor`. That in turn
attempts to call `createAndProcessScript` on its workers, but that
function does not exist anymore.

This shows this was inactive the entire time and I've been thinking for
a long time that this _optimizing_ step doesn't buy much (even while it
was active) and removing it simplified the snapshot-doctor a lot.

However if this behavior is desired then the following should be done:

- revert this commit
- implement the missing function on the worker (history should hold a
  version at some point)
- fix the optimization code to actually perform something (my guess is
  that the `filter` function (`./src/doctor/snapshot-doctor.ts:258`)
  filters out all potential modules to be optimized